Refactor chat bot system and add new commands
Replaces ChatBotConfig with direct input fields, updates ChatBotClient to handle connection/disconnection and user pings, and refactors ChatBotGame to use a command dictionary for faster lookups. Adds AutoHelloResponse, CheckBalanceCommand, and RequestGptCommand, and updates the UI with new buttons and larger font sizes. Removes PrintToChatSignal in favor of PrintToLocalChatSignal, and introduces web request scripts for player data. Scene and asset files updated to reflect these changes.

+++ Assets/Scripts/ChatBot/PlayerObject.cs
@@ -7,7 +7,7 @@ namespace ChatBot
     public class PlayerObject
     {
         public string twitchName;
-        public float gold;
+        public int gold;
 
         public PlayerObject(string twitchName)
         {
